Remove actions - Remove role
