+++ auth/models/role.py
@@ -1,5 +1,7 @@
 # coding: utf-8
 from datetime import datetime
+from sqlalchemy.exc import IntegrityError
+from auth.exceptions import InvalidRoleName, RoleAlreadyExist, RoleNotFound
 from auth.models import Model, db
 
 
@@ -15,3 +17,42 @@ class Role(Model):
     @property
     def is_active(self):
         return self.active
+
+    @classmethod
+    def create(cls, name, description=None):
+        if len(name) < 3:
+            raise InvalidRoleName
+
+        try:
+            role = Role()
+            role.name = name
+            role.description = description
+            role.save()
+            db.session.commit()
+            return role
+        except IntegrityError:
+            raise RoleAlreadyExist
+
+    def edit(self, name=None, description=None):
+        try:
+            if name:
+                if len(name) < 3:
+                    raise InvalidRoleName
+                self.name = name
+            if description:
+                self.description = description
+            self.save()
+            db.session.commit()
+        except IntegrityError:
+            raise RoleAlreadyExist
+        return self
+
+    @classmethod
+    def search_role(cls, name, exactly=False):
+        if exactly:
+            role = cls.query.filter(Role.name == name).all()
+        else:
+            role = cls.query.filter(Role.name.contains(name)).all()
+        if not role:
+            raise RoleNotFound
+        return role
